@@ -15,7 +15,7 @@ export class MemberEditResolver implements Resolve<User> {
     resolve(route: ActivatedRouteSnapshot): Observable<User> {
         return this.userService.getUser(this.authService.docodedToken.nameid).pipe(
             catchError(error => {
-                this.alertify.error('Prolem retrieving your data');
+                this.alertify.error('Problem retrieving your data');
                 this.router.navigate(['/members']);
                 return of(null);
             })
