@@ -87,7 +87,7 @@ namespace DatingApp.API.Controllers
             if (await _repository.SaveAll())
             {
                 var messageToReturn = _mapper.Map<MessageToReturnDto>(message);
-                return CreatedAtRoute("GetMessage", new { id = message.Id}, messageToReturn);
+                return CreatedAtRoute("GetMessage", new { userId, id = message.Id}, messageToReturn);
             }
             throw new Exception("Failed to create message");
         }
