@@ -79,7 +79,7 @@ namespace DatingApp.API.Controllers
             if (await _repository.SaveAll())
             {
                 var photoToReturn = _mapper.Map<PhotoForReturnDto>(photo);
-                return CreatedAtRoute("GetPhoto", new { id = photo.Id}, photoToReturn);
+                return CreatedAtRoute("GetPhoto", new { userId, id = photo.Id}, photoToReturn);
             }
 
             return BadRequest("Could not add the photo");
