@@ -1,14 +1,10 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.IO;
-using System.Linq;
-using System.Threading.Tasks;
 using DatingApp.API.Data;
 using Microsoft.AspNetCore;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.EntityFrameworkCore;
-using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
 
 namespace DatingApp.API
@@ -17,7 +13,7 @@ namespace DatingApp.API
     {
         public static void Main(string[] args)
         {
-            var host = CreateWebHostBuilder(args).Build(); 
+            var host = CreateHostBuilder(args).Build(); 
             /*Seed (create/migrate) database*/
             using(var scope = host.Services.CreateScope()) /*Can't use injection, use services*/
             {
@@ -37,8 +33,11 @@ namespace DatingApp.API
             host.Run();
         }
 
-        public static IWebHostBuilder CreateWebHostBuilder(string[] args) =>
-            WebHost.CreateDefaultBuilder(args)
-                .UseStartup<Startup>();
+        public static IHostBuilder CreateHostBuilder(string[] args) =>
+            Host.CreateDefaultBuilder(args).ConfigureWebHostDefaults(webBuilder => 
+            {
+                webBuilder.UseStartup<Startup>();
+            });
+                
     }
 }
